fix: no ethereum prefix for custom signers (#194)
Originally when `Signer` interface was called as an argument was pass to it data that were already prefixed with the Ethereum prefix (`\x19Ethereum Signed Message:\n${data.length}`). This prevents the use of the main signing methods in Ethereum wallets, which do this prefixing on their own in order to protect users from phishing attacks. 

This patch removes this hardcoded prefixing and leaves it up to the `Signer` implementator to ensure the prefix is applied.